Refactoring: extract TreeOptimizer from SyntaxAnalyzer
