+++ src/Interpreters/ArithmeticOperationsInAgrFuncOptimize.h
@@ -6,6 +6,8 @@
 namespace DB
 {
 
+class ASTFunction;
+
 /// Extract constant arguments out of aggregate functions from child functions
 /// 'sum(a * 2)' -> 'sum(a) * 2'
 /// Rewrites:   sum([multiply|divide]) -> [multiply|divide](sum)
