+++ src/Interpreters/GroupByFunctionKeysVisitor.h
@@ -8,6 +8,7 @@
 #include <Parsers/ASTSelectQuery.h>
 #include <Parsers/ASTSetQuery.h>
 #include <Parsers/ASTTablesInSelectQuery.h>
+#include <Parsers/ASTIdentifier.h>
 #include <Parsers/IAST.h>
 #include <Common/typeid_cast.h>
